% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{county_stat}
\alias{county_stat}
\alias{county_median_age}
\alias{county_race_gender}
\alias{county_educ}
\title{County-level statistics, 1999-2015}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 54791 rows and 16 columns.}
\usage{
county_stat

county_median_age

county_race_gender

county_educ
}
\description{
Variables are downloaded via \code{tidycensus}.

The dataset \code{county_stat} combines \code{county_median_age}, \code{county_race_gender},
and \code{county_educ}.

Note that variables in \code{county_educ} refer to county population 25 years and
over.
}
\keyword{datasets}
